Crud Ingrediente.

+++ pizzaria/src/main/java/br/com/pizzaria/model/SubstituicaoIngrediente.java
@@ -5,9 +5,7 @@ import java.util.Objects;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 /**
@@ -21,7 +19,7 @@ import javax.persistence.Table;
 public class SubstituicaoIngrediente{
 	
 	@Id
-	private Integer id;
+	private Long id;
 	
 	@ManyToOne
 	private Ingrediente ingredienteSubstituido;
@@ -33,12 +31,11 @@ public class SubstituicaoIngrediente{
 	@JoinColumn(name="item_pedido_id", nullable = false	)	
 	private ItemPedido itemPedido;
 	
-
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
